Moved Hasher in proper package
