+++ server/src/main/kotlin/com/kamelia/jellyfish/core/Hasher.kt
@@ -1,4 +1,4 @@
-package com.kamelia.jellyfish.util
+package com.kamelia.jellyfish.core
 
 import at.favre.lib.crypto.bcrypt.BCrypt
 import at.favre.lib.crypto.bcrypt.LongPasswordStrategies
